requirements.txt
requests>=2.31.0
fastmcp>=0.1.0 